Code/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
analytics-python==1.4.0
anyio==3.6.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.22.1
azure-storage-blob==12.11.0
Babel==2.10.3
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.27
botocore==1.27.27
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
determined==0.19.1
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.7.1
flatbuffers==1.12
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.7.1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.48.0
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.3
imageio==2.19.5
ipykernel==6.15.1
ipympl==0.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itk-core==5.3rc4.post1
itk-elastix==0.14.1
itk-filtering==5.3rc4.post1
itk-io==5.3rc4.post1
itk-numerics==5.3rc4.post1
itk-registration==5.3rc4.post1
itk-segmentation==5.3rc4.post1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.6
lmdb==1.3.0
lomond==0.3.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
msrest==0.6.21
multidict==6.0.2
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
neurite==0.2
nibabel==4.0.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pyelastix==1.2
Pygments==2.12.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pystrum==0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.6.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.2.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2022.8.3
tinycss2==1.1.1
torch==1.12.0
torch-similarity==1.0.0
torchaudio==0.12.0
torchmetrics==0.9.2
torchsummary==1.5.1
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
voxelmorph==0.2
Wand==0.6.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.3
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wrapt==1.14.1
yarl==1.7.2
yogadl==0.1.4
